requirements.txt
apache-airflow-providers-amazon==7.2.1
apache-airflow-providers-slack==7.2.0
airflow_multi_dagrun==2.3.1
apache-airflow-providers-databricks==6.2.0
parameterized==0.9.0
